+++ lib/karafka/worker.rb
@@ -6,6 +6,7 @@ module Karafka
     # @param args [Array] controller params and controller topic
     # @note Arguments are provided in Karafka::BaseController enqueue
     def execute(*args)
+      Karafka::App.logger.info("#{self.class}#execute for #{args}")
       self.args = args
       controller.perform
     end
@@ -15,8 +16,13 @@ module Karafka
     # @param args [Array] controller params and controller topic
     def after_failure(*args)
       self.args = args
-      return unless controller.respond_to?(:after_failure)
 
+      unless controller.respond_to?(:after_failure)
+        Karafka::App.logger.warn("#{self.class}#after_failure controller missing for #{args}")
+        return
+      end
+
+      Karafka::App.logger.warn("#{self.class}#after_failure for #{args}")
       controller.after_failure
     end