+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Karafka framework changelog
 
+## 0.1.2
+
+- Logging to log/environment.log
+- Karafka::Runner
+
 ## 0.1.1
 
 - README updates